Move SharedAbortingJoinHandle to be interal for iroh::node::Node
This can't easily be moved off this because Node is Clone *and* awaits
the task.  It's not ideal.
